refactor(core,console,schemas): remove logDto
remove logDto
