+++ packages/core/src/routes/interaction/actions/submit-interaction.ts
@@ -5,6 +5,7 @@ import { conditional } from '@silverhand/essentials';
 import { getLogtoConnectorById } from '#src/libraries/connector.js';
 import { assignInteractionResults } from '#src/libraries/session.js';
 import { encryptUserPassword } from '#src/libraries/user.js';
+import type { LogEntry } from '#src/middleware/koa-audit-log.js';
 import type TenantContext from '#src/tenants/TenantContext.js';
 
 import type { WithInteractionDetailsContext } from '../middleware/koa-interaction-details.js';
@@ -129,9 +130,11 @@ const parseUserProfile = async (
 export default async function submitInteraction(
   interaction: VerifiedInteractionResult,
   ctx: WithInteractionDetailsContext,
-  { provider, libraries, queries }: TenantContext
+  { provider, libraries, queries }: TenantContext,
+  log?: LogEntry
 ) {
   const { hasActiveUsers, findUserById, updateUserById } = queries.users;
+
   const {
     users: { generateUserId, insertUser },
   } = libraries;
@@ -155,10 +158,13 @@ export default async function submitInteraction(
 
     await assignInteractionResults(ctx, provider, { login: { accountId: id } });
 
+    log?.append({ userId: id });
+
     return;
   }
 
   const { accountId } = interaction;
+  log?.append({ userId: accountId });
 
   if (event === InteractionEvent.SignIn) {
     const user = await findUserById(accountId);
